refactor: Move Avatar's models to their own "models" package
