+++ Avatar/src/main/java/com/infomaniak/core/avatar/models/AvatarType.kt
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.infomaniak.core.avatar
+package com.infomaniak.core.avatar.models
 
 import androidx.annotation.DrawableRes
 import coil3.ImageLoader
